Split graphql resolvers in folders and files
